sandbox/sail: Updated portaudio installation script.
This patch changes the way src/common/pa_ringbuffer.o file is added to the
portaudio library. Original script added this file to the COMMON_OBJS list. However, on some platforms this file is automatically added to the OTHER_OBJS
list and as a result pa_ringbuffer.o object is added twice to the portaudio
library. Linking against portaudio library from a piece of code using the
pa_ringbuffer results in linker errors.

git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/sail@3701 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8