BlockAffineComponent now passes tests. Previous problem was a typo in BlockAffineComponent::DotProduct.
Also: Change input parameters of RepeatedAffineComponent tests.
num_repeats=input_dim=output_dim causes tests to fail. That is,
the specific case, where each block is a 1-by-1 matrix.

I imagine this is because of the tests cases themselves, but am
not sure, since I don't fully understand them. Nevertheless, it's
probably not a common case to worry about.