Make CWD check less strict
It might be that `configure` is still invoked from 'src' directory, yet the
directory name itself isn't 'src'. While a real procedure of comparing two
filesystem paths to be equal is quite complex, this small trick does exactly
what we want to do - check that `configure` is invoked from the same directory
where it is.