reimpl log softmax
New: For CuMatrix::LogSoftmax<float>, for dim = 16, speed was 0.0138019 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 16, speed was 0.0133804 gigaflops.
New: For CuMatrix::LogSoftmax<float>, for dim = 32, speed was 0.056202 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 32, speed was 0.052121 gigaflops.
New: For CuMatrix::LogSoftmax<float>, for dim = 64, speed was 0.227829 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 64, speed was 0.186255 gigaflops.
New: For CuMatrix::LogSoftmax<float>, for dim = 128, speed was 0.65638 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 128, speed was 0.65072 gigaflops.
New: For CuMatrix::LogSoftmax<float>, for dim = 256, speed was 2.15268 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 256, speed was 1.64888 gigaflops.
New: For CuMatrix::LogSoftmax<float>, for dim = 512, speed was 5.1179 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 512, speed was 3.85136 gigaflops.
New: For CuMatrix::LogSoftmax<float>, for dim = 1024, speed was 10.8209 gigaflops.
Old: For CuMatrix::LogSoftmax<float>, for dim = 1024, speed was 6.76963 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 16, speed was 0.0133584 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 16, speed was 0.011373 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 32, speed was 0.0533796 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 32, speed was 0.0528196 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 64, speed was 0.202721 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 64, speed was 0.170107 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 128, speed was 0.627234 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 128, speed was 0.722198 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 256, speed was 1.89987 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 256, speed was 1.44478 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 512, speed was 4.14807 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 512, speed was 3.37973 gigaflops.
New: For CuMatrix::LogSoftmax<double>, for dim = 1024, speed was 6.70849 gigaflops.
Old: For CuMatrix::LogSoftmax<double>, for dim = 1024, speed was 4.96657 gigaflops.