Max change per component (#918)
Add the capability to enforce a maximum change per component in nnet3, at the minibatch level.
This CHANGES THE DEFAULT BEHAVIOR OF EXISTING SCRIPTS.
nnet3 and nnet3+chain config-generation scripts will now by default apply a max-change at the per component level.  You can override this (for back compatibility) by adding --max-change-per-component=0.0 --max-change-per-component-final=0.0 to the make_configs.py scripts.  However, we don't recommend doing this unless it is part of an existing experimental setup that you need to keep consistent, because we found that the max-change-per-component is slightly helpful (0.1% or 0.2% absolute in our experiments).